backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.5.0
